git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@16822 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/sixel.c
@@ -712,7 +712,7 @@ static MagickBooleanType sixel_encode_impl(unsigned char *pixels, size_t width,s
         return (MagickFalse);
     }
     sixel_advance(context, nwrite);
-    nwrite = sprintf((char *)context->buffer + context->pos, "\"1;1;%d;%d", width, height);
+    nwrite = sprintf((char *)context->buffer + context->pos, "\"1;1;%d;%d", (int) width, (int) height);
     if (nwrite <= 0) {
         RelinquishNodesAndMap;
         return (MagickFalse);
